+++ packages/frontend/src/App.tsx
@@ -6,7 +6,19 @@ import LoginPage from './pages/LoginPage'
 import SignUpPage from './pages/SignUpPage'
 import ChatRoomPage from './pages/ChatRoomPage'
 import { Container } from 'react-bootstrap'
+import axios from 'axios'
 
+axios.defaults.baseURL = process.env.REACT_APP_CHAT_API || 'http://localhost:4000'
+axios.interceptors.request.use((config) => {
+  if (!config?.headers) {
+    config.headers = {}
+  }
+  const jwt = localStorage.getItem('chatapp')
+  if (jwt) {
+    config.headers['authorization'] = `Bearer ${jwt}`
+  }
+  return config
+})
 function App() {
   return (
     <div className='App'>
