+++ packages/frontend/src/pages/LoginPage.tsx
@@ -20,7 +20,7 @@ const LoginPage = () => {
         setMessage('')
       }, 5000)
     }
-    const loginResponse = await axios.post(`${process.env.REACT_APP_CHAT_API}/api/users/login`, {
+    const loginResponse = await axios.post('/api/users/login', {
       username: username,
       password: password,
     })
